add member selector
